move adaptive predicate pushdown test
